fix modal
